requirements.txt
alembic==1.4.2
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.0
marshmallow-sqlalchemy==0.23.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.18
uuid==1.30
Werkzeug==1.0.1
